Modified detection annote fields
Modified detection of the annote field so that the search algorithm is not prematurely closed if there is a comma after a closing curly brace within the text of the annotation. The algorithm now looks for a newline after the closing curly brace and comma, which is less likely to have false positives.